+++ CHANGELOG.md
@@ -63,3 +63,4 @@
 
 ## 13.2.1
 * fix accessibility in input-text, input-dates, paginator
+* fix structure of autocomplete-obj and autocomplete-strg